fix style for rounded buttons
