+++ src/app/screens/Login/index.tsx
@@ -3,8 +3,9 @@ import { ImageBackground, Image, TextInput } from 'react-native';
 import loginBackground from '@assets/bc_inicio.png';
 import appLogo from '@assets/wbooks_logo.png';
 import Button from '@components/Button';
+import { FIELD } from '@constants/styleHelpers';
 
-import styles, { FIELD } from './styles';
+import styles from './styles';
 
 function Login() {
   return (
